Remove DetailView when graphView is moved or zoomed
