+++ src/main/java/tudelft/ti2806/pl3/controls/KeyController.java
@@ -92,6 +92,8 @@ public class KeyController implements KeyListener {
 			float newViewCenter = oldViewCenter - move;
 			graphController.moveView(newViewCenter);
 		}
+
+		app.getGraphController().removeDetailView();
 	}
 	
 	/**
