+++ src/main/java/tudelft/ti2806/pl3/visualization/GraphController.java
@@ -2,6 +2,7 @@ package tudelft.ti2806.pl3.visualization;
 
 import tudelft.ti2806.pl3.Controller;
 import tudelft.ti2806.pl3.data.filter.Filter;
+import tudelft.ti2806.pl3.data.gene.Gene;
 import tudelft.ti2806.pl3.data.graph.DataNode;
 import tudelft.ti2806.pl3.exception.NodeNotFoundException;
 
@@ -89,7 +90,11 @@ public class GraphController implements Controller {
 		return graphView.getZoomCenter();
 	}
 
-	public void centerOnNode(DataNode node) throws NodeNotFoundException {
-		graphView.centerOnNode(node);
+	public void centerOnNode(DataNode node, Gene selected) throws NodeNotFoundException {
+		graphView.centerOnNode(node, selected);
+	}
+
+	public void removeDetailView() {
+		graphView.removeDetailView();
 	}
 }
